==> jupyter_notebooks/01 - DataCollection.ipynb <==
# **Data Collection Notebook**

## Objectives

* Fetch data from Kaggle and save it as raw data.
* Inspect the data and save it under outputs/datasets/collection

## Inputs

*   Kaggle JSON file - the authentication token.

## Outputs

* Generate Dataset: outputs/datasets/collection/TelcoCustomerChurn.csv


---

# Install python packages in the notebooks

%pip install -r /Users/Endeavour/Code/customer-churn-predictor/customer-churn-predictor/requirements.txt

%pip install matplotlib --no-deps
%pip install streamlit --no-deps

# Change Working Directory

import os
current_dir = os.getcwd()
current_dir

Set the parent of the current directory as the new current directory.
* os.path.dirname() gets the parent directory
* os.chir() defines the new current directory

os.chdir(os.path.dirname(current_dir))
print("You set a new current directory")

Confirm the new current directory

current_dir = os.getcwd()
current_dir

# Download Dataset

%pip install kaggle==1.5.12

import os
os.environ['KAGGLE_CONFIG_DIR'] = os.getcwd()
! chmod 600 kaggle.json

Download and assign the dataset a path.

KaggleDatasetPath = "gyanshashwat1611/telecom-churn-dataset"
DestinationFolder = "inputs/datasets/raw"   
! kaggle datasets download -d {KaggleDatasetPath} -p {DestinationFolder}

Unzip the downloaded file, delete the zip file and delete the kaggle.json file

! unzip {DestinationFolder}/*.zip -d {DestinationFolder} \
  && rm {DestinationFolder}/*.zip \
  && rm kaggle.json

---

# Load and Inspect Data

import pandas as pd
df = pd.read_csv(f"/Users/Endeavour/Code/customer-churn-predictor/inputs/datasets/raw/WA_Fn-UseC_-Telco-Customer-Churn.csv")
df.head()

**DataFrame Summary:**

df.info()

Confirm if there are duplicated `customerID`:

df[df.duplicated(subset=['customerID'])]

Converting `TotalCharges` to numeric

df['TotalCharges'] = pd.to_numeric(df['TotalCharges'] ,errors='coerce')

Check `TotalCharges` data type

df['TotalCharges'].dtype

Currently, `Churn` is a categorical variable: Yes or No. We will replace/convert it to an integer as the ML model requires numeric variables. 

df['Churn'].unique()

df['Churn'] = df['Churn'].replace({"Yes":1, "No":0})

Check the `Churn` data type.

df['Churn'].dtype

# Push files to Repo

import os
try:
  os.makedirs(name='/Users/Endeavour/Code/customer-churn-predictor/outputs/datasets/collection') # create outputs/datasets/collection folder
except Exception as e:
  print(e)

df.to_csv(f"/Users/Endeavour/Code/customer-churn-predictor/outputs/datasets/collection/TelcoCustomerChurn.csv",index=False)
