==> jupyter_notebooks/02 - ChurnedCustomerStudy.ipynb <==
# Churned Customer Study Notebook

## Objectives

*   Answer business requirement 1: 
    * The client is interested to understand the patterns from the customer base, so the client can learn the most relevant variables that are correlated to a churned customer.

## Inputs

* outputs/datasets/collection/TelcoCustomerChurn.csv

## Outputs

* generate code that answers business requirement 1 and can be used to build the Streamlit App






---

# Change working directory

import os
current_dir = os.getcwd()
current_dir

We want to make the parent of the current directory the new current directory
* os.path.dirname() gets the parent directory
* os.chir() defines the new current directory

os.chdir(os.path.dirname(current_dir))
print("You set a new current directory")

Confirm the new current directory

current_dir = os.getcwd()
current_dir

---

# Load Data

The customer ID variable is not useful for the analysis.

import pandas as pd
df = (pd.read_csv("/Users/Endeavour/Code/customer-churn-predictor/customer-churn-predictor/outputs/datasets/collection/TelcoCustomerChurn.csv")
    .drop(['customerID'], axis=1)
    )
df.head(10)

# Data Exploration

import sweetviz as sv
import pandas as pd

df = pd.read_csv("/Users/Endeavour/Code/customer-churn-predictor/customer-churn-predictor/outputs/datasets/collection/TelcoCustomerChurn.csv")
df = df.drop(['customerID'], axis=1)

report = sv.analyze(df)
report.show_notebook()

# Correlation Study

Here the categorical data is encoded using OneHotEncoder to a binary format that can be enterpreted by the ML algorithm.

from feature_engine.encoding import OneHotEncoder
encoder = OneHotEncoder(variables=df.columns[df.dtypes=='object'].to_list(), drop_last=False)
df_ohe = encoder.fit_transform(df)
print(df_ohe.shape)
df_ohe.head(10)

Use `.corr()` for `spearman` and `pearson` methods, and investigate the correlation of churn to the top 10 alternative factors.

* 'pearson' is a method that calculates the correlation coefficient between two variables. It measures the linear relationship between two variables.
* 'spearman' is a method that calculates the correlation coefficient between two variables. It measures the monotonic relationship 
between two variables.

* Sort values considering the absolute value, by setting `key=abs`

corr_spearman = df_ohe.corr(method='spearman')['Churn'].sort_values(key=abs, ascending=False)[1:].head(10)
corr_spearman

corr_pearson = df_ohe.corr(method='pearson')['Churn'].sort_values(key=abs, ascending=False)[1:].head(10)
corr_pearson

For both methods, we notice weak or moderate levels of correlation between Churn and a given variable. 
* Ideally, we pursue strong correlation levels. However, this is not always possible.

The top five correlation levels at `df_ohe` will be considered, then study the associated variables at `df`

top_n = 5
set(corr_pearson[:top_n].index.to_list() + corr_spearman[:top_n].index.to_list())

Therefore we are studying at df the following variables. We will investigate if:
* A churned customer typically has a month-to-month contract.
* A churned customer typically has fiber optic.
* A churned customer typically doesn't have tech support.
* A churned customer doesn't have online security.
* A churned customer typically has low tenure levels.

vars_to_study = ['Contract', 'InternetService', 'OnlineSecurity', 'TechSupport', 'tenure']
vars_to_study

# Exploratory Data Analysis on selected variables

df_eda = df.filter(vars_to_study + ['Churn'])
print(df_eda.shape)
df_eda.head(3)

## Variables Distribution by Churn

Here are function to loop over the variables and plot the distribution (numerical and categorical) coloured by Churn.
If the variable is numerical, we plot the distribution using a histogram.
If the variable is categorical, we plot the distribution using a bar chart.

import matplotlib.pyplot as plt
import seaborn as sns
sns.set_style('whitegrid')


def plot_categorical(df, col, target_var):

    plt.figure(figsize=(12, 5))
    sns.countplot(data=df, x=col, hue=target_var, order=df[col].value_counts().index)
    plt.xticks(rotation=90)
    plt.title(f"{col}", fontsize=20, y=1.05)
    plt.show()


def plot_numerical(df, col, target_var):
    plt.figure(figsize=(8, 5))
    sns.histplot(data=df, x=col, hue=target_var, kde=True, element="step")
    plt.title(f"{col}", fontsize=20, y=1.05)
    plt.show()


target_var = 'Churn'
for col in vars_to_study:
    if df_eda[col].dtype == 'object':
        plot_categorical(df_eda, col, target_var)
        print("\n\n")
    else:
        plot_numerical(df_eda, col, target_var)
        print("\n\n")


**General Trends:**

* Tenure: Churned customers tend to have lower tenure levels.
* Contract: Churned customers tend to have month-to-month contracts.
* Internet Service: Churned customers tend to have fiber optic internet service.
* Online Security: Churned customers tend to not have online security.
* Tech Support: Churned customers tend to not have tech support.


---

## Parallel Plot

Create a separate DataFrame and transform the neumerical data of `tenure` into categorical bins (6 month bins) for visualising at `parallel_categories()` plot.


from feature_engine.discretisation import ArbitraryDiscretiser
import numpy as np
tenure_map = [-np.Inf, 6, 12, 18, 24, np.Inf]
disc = ArbitraryDiscretiser(binning_dict={'tenure': tenure_map})
df_parallel = disc.fit_transform(df_eda)
df_parallel.head()

disc.binner_dict_['tenure']

Create a map to replace the `tenure` variable with more informative levels.

n_classes = len(tenure_map) - 1
classes_ranges = disc.binner_dict_['tenure'][1:-1]

labels_map = {}
for n in range(0, n_classes):
    if n == 0:
        labels_map[n] = f"<{classes_ranges[0]}"
    elif n == n_classes-1:
        labels_map[n] = f"+{classes_ranges[-1]}"
    else:
        labels_map[n] = f"{classes_ranges[n-1]} to {classes_ranges[n]}"

labels_map


Replace using `.replace()`

df_parallel['tenure'] = df_parallel['tenure'].replace(labels_map)
df_parallel.head()

Creates multi-dimensional categorical data plot

import plotly.express as px
import plotly.graph_objects as go

fig = px.parallel_categories(df_parallel, color="Churn")

# Convert to FigureWidget
fig_widget = go.FigureWidget(fig)

# Display the widget
fig_widget

---

# Conclusions and Next steps

The correlations and plots interpretation converge. 
* A churned customer typically has a month-to-month contract.
* A churned customer typically has fiber optic.
* A churned customer typically doesn't have tech support.
* A churned customer doesn't have online security.
* A churned customer typically has low tenure levels.

---
